:art::recycle:refactor code
