+++ anno2xlsx.go
@@ -2,17 +2,13 @@ package main
 
 import (
 	"flag"
-	"fmt"
 	"log"
 	_ "net/http/pprof"
 	"os"
 	"runtime/pprof"
-	"strings"
-	"time"
 
 	"github.com/liserjrqlxue/goUtil/osUtil"
 	"github.com/liserjrqlxue/goUtil/simpleUtil"
-	"github.com/liserjrqlxue/goUtil/textUtil"
 	"github.com/tealeg/xlsx/v3"
 )
 
@@ -49,50 +45,6 @@ func init() {
 	parseToml()
 	parseCfg()
 
-}
-
-func parseList() {
-	sampleList = strings.Split(*list, ",")
-	for _, sample := range sampleList {
-		sampleMap[sample] = true
-		quality := make(map[string]string)
-		quality["样本编号"] = sample
-		qualitys = append(qualitys, quality)
-	}
-}
-
-func parseQC() {
-	var karyotypeMap = make(map[string]string)
-	if *karyotype != "" {
-		karyotypeMap, err = textUtil.Files2Map(*karyotype, "\t", true)
-		simpleUtil.CheckErr(err)
-	}
-	// load coverage.report
-	if *qc != "" {
-		loadQC(*qc, *kinship, qualitys, *wgs)
-		for _, quality := range qualitys {
-			for k, v := range qualityKeyMap {
-				quality[k] = quality[v]
-			}
-			quality["核型预测"] = karyotypeMap[quality["样本编号"]]
-			if *wesim {
-				var qcArray []string
-				for _, key := range qcColumn {
-					qcArray = append(qcArray, quality[key])
-				}
-				_, err = fmt.Fprintln(qcFile, strings.Join(qcArray, "\t"))
-				simpleUtil.CheckErr(err)
-			}
-		}
-		if *wesim {
-			simpleUtil.CheckErr(qcFile.Close())
-		}
-
-		ts = append(ts, time.Now())
-		step++
-		logTime(ts, step-1, step, "load coverage.report")
-		loadFilterStat(*filterStat, qualitys[0])
-	}
 }
 
 func main() {
